kode/requirements.txt
pandas
snowflake-sqlalchemy
sqlalchemy
kafka-python